+++ utils/gridadmin.py
@@ -13,6 +13,8 @@ ogr.UseExceptions()  # fail fast
 
 SPATIALITE_DRIVER_NAME = 'SQLite'
 
+from ..utils.user_messages import log
+
 
 class QgisNodesOgrExporter(BaseOgrExporter):
     """
@@ -56,7 +58,9 @@ class QgisNodesOgrExporter(BaseOgrExporter):
             SPATIALITE_DRIVER_NAME,
         }
 
-    def save(self, file_name, node_data, target_epsg_code, **kwargs):
+    def save(
+            self, file_name, layer_name, node_data, target_epsg_code,
+            **kwargs):
         """
         save to file format specified by the driver, e.g. shapefile
 
@@ -65,11 +69,14 @@ class QgisNodesOgrExporter(BaseOgrExporter):
         """
         assert self.driver is not None
         sr = get_spatial_reference(target_epsg_code)
-        self.del_datasource(file_name)
-        data_source = self.driver.CreateDataSource(
-            file_name, ["SPATIALITE=YES"])
+        if not os.path.exists(file_name):
+            data_source = self.driver.CreateDataSource(
+                file_name, ["SPATIALITE=YES"])
+        else:
+            data_source = self.driver.Open(file_name, update=1)
+
         layer = data_source.CreateLayer(
-            str(os.path.splitext(os.path.basename(file_name))[0]),
+            layer_name,
             sr,
             geom_type=ogr.wkbPoint,
             options=['FORMAT=SPATIALITE']
@@ -113,6 +120,7 @@ class QgisNodesOgrExporter(BaseOgrExporter):
                 feature.SetFID(node_data['id'][i])
             layer.CreateFeature(feature)
             feature.Destroy()
+        data_source = None
 
 
 class QgisKCUDescriptor(KCUDescriptor):
@@ -209,7 +217,9 @@ class QgisLinesOgrExporter(BaseOgrExporter):
         }
         self.driver = None
 
-    def save(self, file_name, line_data, target_epsg_code, **kwargs):
+    def save(
+            self, file_name, layer_name, line_data, target_epsg_code,
+            **kwargs):
         """
         save to file format specified by the driver, e.g. shapefile
 
@@ -220,12 +230,14 @@ class QgisLinesOgrExporter(BaseOgrExporter):
 
         kcu_dict = QgisKCUDescriptor()
         sr = get_spatial_reference(target_epsg_code)
+        if not os.path.exists(file_name):
+            data_source = self.driver.CreateDataSource(
+                file_name, ["SPATIALITE=YES"])
+        else:
+            data_source = self.driver.Open(file_name, update=1)
 
-        self.del_datasource(file_name)
-        data_source = self.driver.CreateDataSource(
-            file_name, ["SPATIALITE=YES"])
         layer = data_source.CreateLayer(
-            str(os.path.splitext(os.path.basename(file_name))[0]),
+            layer_name,
             sr,
             geom_type=ogr.wkbLineString,
             options=['FORMAT=SPATIALITE'],
@@ -282,3 +294,90 @@ class QgisLinesOgrExporter(BaseOgrExporter):
 
             layer.CreateFeature(feature)
             feature.Destroy()
+        data_source = None
+
+
+class QgisPumpsOgrExporter(BaseOgrExporter):
+    """
+    Exports to ogr formats. You need to set the driver explicitly
+    before calling save()
+    """
+    FIELDS = OrderedDict([
+        ('id', 'int'),
+        ('node_idx1', 'int'),
+        ('node_idx2', 'int'),
+    ])
+
+    FIELD_NAME_MAP = OrderedDict([
+        ('id', 'id'),
+        ('node_idx1', 'node1_id'),
+        ('node_idx2', 'node2_id'),
+    ])
+
+    def __init__(self, node_data):
+        self.node_data = node_data
+        self.driver = None
+
+    def save(
+            self, file_name, layer_name, pump_data, target_epsg_code,
+            **kwargs):
+        """
+        save to file format specified by the driver, e.g. shapefile
+
+        :param file_name: name of the outputfile
+        :param line_data: dict of line data
+        """
+        assert self.driver is not None
+
+        sr = get_spatial_reference(target_epsg_code)
+        if not os.path.exists(file_name):
+            data_source = self.driver.CreateDataSource(
+                file_name, ["SPATIALITE=YES"])
+        else:
+            data_source = self.driver.Open(file_name, update=1)
+
+        layer = data_source.CreateLayer(
+            layer_name,
+            sr,
+            geom_type=ogr.wkbLineString,
+            options=['FORMAT=SPATIALITE'],
+        )
+
+        for field_name, field_type in self.FIELDS.iteritems():
+            layer.CreateField(ogr.FieldDefn(
+                    str(field_name), OGR_FIELD_TYPE_MAP[field_type])
+            )
+        _definition = layer.GetLayerDefn()
+
+        for i in xrange(pump_data['id'].size):
+            line = ogr.Geometry(ogr.wkbLineString)
+            node1_id = pump_data['node1_id'][i]
+            node2_id = pump_data['node2_id'][i]
+
+            for node_id in [node1_id, node2_id]:
+                try:
+                    line.AddPoint_2D(
+                        self.node_data['coordinates'][0][node_id],
+                        self.node_data['coordinates'][1][node_id],
+                    )
+                except IndexError:
+                    log("Invalid node id: %s" % node_id)
+
+            feature = ogr.Feature(_definition)
+            feature.SetGeometry(line)
+            for field_name, field_type in self.FIELDS.iteritems():
+                fname = self.FIELD_NAME_MAP[field_name]
+                raw_value = pump_data[fname][i]
+                value = TYPE_FUNC_MAP[field_type](raw_value)
+                feature.SetField(str(field_name), value)
+                # Using ['FID=id'] in CreateLayer doesn't work on GDAL < 2.0,
+                # thus FID defaults to 'OGC_FID', which sucks.
+                # See: http://www.gdal.org/drv_sqlite.html
+                # To circumvent this, we set 'OGC_FID' to our 'id', so we
+                # can do feature.id() in QGIS and get the line index without
+                # having to specify the 'id' column.
+                feature.SetFID(pump_data['id'][i])
+
+            layer.CreateFeature(feature)
+            feature.Destroy()
+        data_source = None